pyproject.toml
[project]
name = "pixelforge-mcp"
version = "0.2.0"
description = "AI-powered image generation MCP server with multi-model support"
authors = [{name = "Ahmed Al-Eryani", email = "ahmed@tehnolabs.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "AGPL-3.0-or-later"}
keywords = ["mcp", "image-generation", "gemini", "ai", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries",
    "Framework :: Pydantic :: 2",
]

dependencies = [
    "fastmcp>=2.0.0,<3.0.0",
    "gemini-imagen>=0.6.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
s3 = [
    "boto3>=1.28.0",
]
image = [
    "pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/tehnolabs/pixelforge-mcp"
Repository = "https://github.com/tehnolabs/pixelforge-mcp"
"Bug Tracker" = "https://github.com/tehnolabs/pixelforge-mcp/issues"

[project.scripts]
pixelforge-mcp = "pixelforge_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310', 'py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
